+++ src/canvas/widgets/mem_graph.rs
@@ -93,10 +93,7 @@ impl MemGraphWidget for Painter {
                 const TITLE_BASE: &str = " Memory ── Esc to go back ";
                 let repeat_num =
                     usize::from(draw_loc.width).saturating_sub(TITLE_BASE.chars().count() + 2);
-                let result_title = format!(
-                    " Memory ─{}─ Esc to go back ",
-                    "─".repeat(repeat_num as usize)
-                );
+                let result_title = format!(" Memory ─{}─ Esc to go back ", "─".repeat(repeat_num));
 
                 result_title
             } else {
