+++ src/canvas/widgets/temp_table.rs
@@ -70,7 +70,7 @@ impl TempTableWidget for Painter {
                     usize::from(draw_loc.width).saturating_sub(TITLE_BASE.chars().count() + 2);
                 let result_title = format!(
                     " Temperatures ─{}─ Esc to go back ",
-                    "─".repeat(repeat_num as usize)
+                    "─".repeat(repeat_num )
                 );
 
                 result_title
